Refactor and add test
