+++ react/src/components/AssetsPanel/FeatureFileTree.tsx
@@ -11,7 +11,7 @@
  - add select highlight and update query param
  - make sure all expaneded
  - right now its 200px wide but previously 250
- - add test for featureCollectionToFileNodeArray and refactor into util?
+ - consider separte task for getting sytem (and systme display name) added to feature-asset
 
 
  */
@@ -21,86 +21,10 @@ import { useTable, useExpanded, Column } from 'react-table';
 
 import Icon from '../../core-components/Icon';
 
-import { FeatureCollection, Feature } from '../../types';
+import { featureCollectionToFileNodeArray } from '../../utils/featureTreeUtils';
+
+import { FeatureCollection } from '../../types';
 import styles from './FeatureFileTree.module.css';
-
-interface FeatureFileNode {
-  id: string;
-  name: string;
-  isDirectory: boolean;
-  children?: FeatureFileNode[];
-}
-
-function createFeatureFileNode(
-  id: string,
-  name: string,
-  isDirectory: boolean,
-  children?: FeatureFileNode[]
-): FeatureFileNode {
-  return { id, name, isDirectory, ...(children && { children }) };
-}
-
-function getFullPathFromFeature(feature: Feature): string {
-  const firstAsset = feature.assets[0];
-  if (firstAsset?.display_path) {
-    return firstAsset.display_path;
-  }
-  if (firstAsset) {
-    return firstAsset.id.toString();
-  }
-  return feature.id.toString();
-}
-
-/**
- * Convert a feature collection to a tree of file nodes.
- *
- * This returns an array of only top-level nodes (which contain
- * a node for all the features + their parent directories)
- */
-function featureCollectionToFileNodeArray(
-  featureCollection: FeatureCollection
-): FeatureFileNode[] {
-  const rootNodes: FeatureFileNode[] = [];
-  const nodeMap: { [key: string]: FeatureFileNode } = {};
-
-  // Sort features to ensure consistent order
-  const sortedFeatures = featureCollection.features.sort((a, b) =>
-    getFullPathFromFeature(a).localeCompare(getFullPathFromFeature(b))
-  );
-
-  sortedFeatures.forEach((feature) => {
-    const nodePath = getFullPathFromFeature(feature);
-    const parts = nodePath.split('/');
-
-    let currentPath = '';
-    let currentNode: FeatureFileNode | null = null;
-
-    parts.forEach((part, index) => {
-      currentPath += (currentPath ? '/' : '') + part;
-      const isLast = index === parts.length - 1;
-
-      if (!nodeMap[currentPath]) {
-        const newNode = createFeatureFileNode(
-          isLast ? feature.id.toString() : currentPath,
-          part,
-          !isLast
-        );
-        nodeMap[currentPath] = newNode;
-
-        if (currentNode) {
-          if (!currentNode.children) currentNode.children = [];
-          currentNode.children.push(newNode);
-        } else {
-          rootNodes.push(newNode);
-        }
-      }
-
-      currentNode = nodeMap[currentPath];
-    });
-  });
-
-  return rootNodes;
-}
 
 interface FeatureFileTreeProps {
   /**
